+++ src/modelo/tablero/tipos_casilleros/RetrocesoDinamico.java
@@ -17,7 +17,7 @@ public class RetrocesoDinamico extends Casillero {
 
         for (int i = 2; i <= 6; i++) sumaObtenida.put(i, dados - unJugador.getCantidadTotalPropiedades());
 
-        for (int i = 7; i <= 10; i++) sumaObtenida.put(i, unJugador.getDinero() % dados);
+        for (int i = 7; i <= 10; i++) sumaObtenida.put(i, (int)unJugador.getDinero() % dados);
 
         for (int i = 11; i <= 12; i++) sumaObtenida.put(i, dados - 2);
